# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HopeCash API
  description: |
    API REST da plataforma HopeCash (NewHope) — gestão financeira pessoal e familiar,
    offline-first. Todas as entidades sincronizáveis carregam `id` (UUID), `version`,
    `updated_at` e `deleted_at` (soft delete). Erros seguem o envelope
    `{ "error": { "code", "message", "details?" } }`.
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /retaguarda/users:
    get:
      tags: [Retaguarda]
      summary: Lista usuários da retaguarda
      responses: { '200': { description: Lista } }
    post:
      tags: [Retaguarda]
      summary: Cria usuário da retaguarda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                role: { type: string, enum: [admin, superuser], default: admin }
      responses: { '201': { description: Criado } }
...
  /retaguarda/users/{id}:
    put:
      tags: [Retaguarda]
      summary: Atualiza nome/papel/status/senha de um usuário da retaguarda
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses: { '200': { description: Atualizado } }
    delete:
      tags: [Retaguarda]
      summary: Exclui (soft delete) um usuário da retaguarda
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses: { '200': { description: OK } }
  /retaguarda/app-users:
    get:
      tags: [Retaguarda]
      summary: Lista/pesquisa usuários do app
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, blocked, pending_deletion] } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses: { '200': { description: 'Página de usuários + meta.total' } }
  /retaguarda/app-users/{id}:
    get:
      tags: [Retaguarda]
      summary: Detalhe de um usuário do app
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses: { '200': { description: Usuário + sessões ativas } }
  /retaguarda/app-users/{id}/status:
    patch:
      tags: [Retaguarda]
      summary: Bloqueia/desbloqueia um usuário do app
      parameters: [{ $ref: '#/components/parameters/id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, blocked] }
      responses: { '200': { description: Usuário atualizado } }
  /retaguarda/app-users/{id}/reset-password:
    post:
      tags: [Retaguarda]
      summary: Gera senha provisória e envia por e-mail ao usuário do app
      description: |
        Redefine a senha do usuário do app para uma senha provisória, encerra as
        sessões ativas e envia o código por e-mail. Quando o envio está desabilitado
        (`MAIL_ENABLED=false`) o código é devolvido no campo `data.code`.
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses:
        '200':
          description: Resultado do reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message: { type: string }
                      email: { type: string }
                      email_sent: { type: boolean }
                      code: { type: string, description: 'Presente apenas com MAIL_ENABLED=false' }

  # Campanhas de notificação push — criação/edição livres para admin e
  # superuser; envio, agendamento, cancelamento e reprocessamento restritos a superuser.
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...